+++ RELEASENOTES.md
@@ -10,6 +10,9 @@
 *   ExoPlayer:
     *   Add `reset` to `BasePreloadManager` to release all the holding sources
         while keep the preload manager instance.
+    *   Add `ExoPlayer.setPriority` (and `Builder.setPriority`) to define the
+        priority value used in `PriorityTaskManager` and for MediaCodec
+        importance from API 35.
 *   Transformer:
     *   Work around a decoder bug where the number of audio channels was capped
         at stereo when handling PCM input.